+++ src/MudBlazor/Components/Chart/Charts/Line.razor.cs
@@ -208,7 +208,7 @@ namespace MudBlazor.Charts
                             gridValueY = verticalStartSpace;
                         }
 
-                        var gridValue = (double)dataLine * verticalSpace / gridYUnits;
+                        var gridValue = dataLine * verticalSpace / gridYUnits;
                         gridValueY = boundHeight - (gridValueY + gridValue);
                         chartLine = chartLine + ToS(gridValueX) + " " + ToS(gridValueY);
                     }
